moved argument types and added wp list
