+++ src/main/java/meteordevelopment/meteorclient/systems/commands/commands/WaypointCommand.java
@@ -5,27 +5,15 @@
 
 package meteordevelopment.meteorclient.systems.commands.commands;
 
-import com.mojang.brigadier.StringReader;
-import com.mojang.brigadier.arguments.ArgumentType;
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
-import com.mojang.brigadier.context.CommandContext;
-import com.mojang.brigadier.exceptions.CommandSyntaxException;
-import com.mojang.brigadier.exceptions.DynamicCommandExceptionType;
-import com.mojang.brigadier.suggestion.Suggestions;
-import com.mojang.brigadier.suggestion.SuggestionsBuilder;
 import meteordevelopment.meteorclient.systems.commands.Command;
+import meteordevelopment.meteorclient.systems.commands.arguments.WaypointArgumentType;
 import meteordevelopment.meteorclient.systems.waypoints.Waypoint;
 import meteordevelopment.meteorclient.systems.waypoints.Waypoints;
 import meteordevelopment.meteorclient.utils.player.PlayerUtils;
 import net.minecraft.command.CommandSource;
-import net.minecraft.text.LiteralText;
 import net.minecraft.util.Formatting;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import java.util.concurrent.CompletableFuture;
 
 import static com.mojang.brigadier.Command.SINGLE_SUCCESS;
 
@@ -36,7 +24,7 @@ public class WaypointCommand extends Command {
 
     @Override
     public void build(LiteralArgumentBuilder<CommandSource> builder) {
-        builder.then(literal("get").executes(ctx -> {
+        builder.then(literal("list").executes(context -> {
             if (Waypoints.get().waypoints.isEmpty()) error("No created waypoints.");
             else {
                 info(Formatting.WHITE + "Created Waypoints:");
@@ -45,7 +33,9 @@ public class WaypointCommand extends Command {
                 }
             }
             return SINGLE_SUCCESS;
-        }).then(argument("waypoint", WaypointArgumentType.waypoint()).executes(context -> {
+        }));
+
+        builder.then(literal("get").then(argument("waypoint", WaypointArgumentType.waypoint()).executes(context -> {
             Waypoint waypoint = WaypointArgumentType.getWaypoint(context, "waypoint");
             info("Name: " + Formatting.WHITE + waypoint.name);
             info("Actual Dimension: " + Formatting.WHITE + waypoint.actualDimension);
@@ -96,37 +86,6 @@ public class WaypointCommand extends Command {
             Waypoints.get().save();
             return SINGLE_SUCCESS;
         })));
-    }
-
-    public static class WaypointArgumentType implements ArgumentType<String> {
-        private static final DynamicCommandExceptionType NO_SUCH_WAYPOINT = new DynamicCommandExceptionType(name -> new LiteralText("Waypoint with name '" + name + "' doesn't exist."));
-
-        public static WaypointArgumentType waypoint() {return new WaypointArgumentType();}
-
-        public static Waypoint getWaypoint(final CommandContext<?> context, final String name) {
-            return Waypoints.get().get(context.getArgument(name, String.class));
-        }
-
-        @Override
-        public String parse(StringReader reader) throws CommandSyntaxException {
-            final String argument = reader.getRemaining();
-            reader.setCursor(reader.getTotalLength());          //ty sb :pray:
-
-            if (Waypoints.get().get(argument) == null) throw NO_SUCH_WAYPOINT.create(argument);
-            return argument;
-        }
-
-        @Override
-        public <S> CompletableFuture<Suggestions> listSuggestions(CommandContext<S> context, SuggestionsBuilder builder) {
-            return CommandSource.suggestMatching(getExamples(), builder);
-        }
-
-        @Override
-        public Collection<String> getExamples() {
-            List<String> names = new ArrayList<>();
-            for (Waypoint waypoint : Waypoints.get()) names.add(waypoint.name);
-            return names;
-        }
     }
 
     private String waypointPos(Waypoint waypoint) {
